requirements.txt
aiohttp==3.8.6
aioresponses==0.7.4
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
chromedriver-binary-auto==0.3.1
Django==4.2.7
django-environ==0.11.2
djangorestframework==3.14.0
dnspython==2.4.2
frozenlist==1.4.0
future==0.18.3
h11==0.14.0
httpretty==1.1.4
idna==3.4
lxml==4.9.3
multidict==6.0.4
outcome==1.3.0.post0
Pillow==10.1.0
pycparser==2.21
pymemcache==4.0.0
PySocks==1.7.1
python-Wappalyzer==0.3.1
python-whois==0.8.0
pytz==2023.3.post1
requests==2.31.0
selenium==4.15.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
trio==0.23.1
trio-websocket==0.11.1
tzdata==2023.3
urllib3==2.0.7
wsproto==1.2.0
yarl==1.9.2
